requirements.txt
flask==1.1.1
flask_restplus==0.13.0
werkzeug==0.16.1
psycopg2==2.9.1
pyodbc==4.0.32
Flask-APScheduler==1.12.2
requests==2.26.0
python-dotenv==0.19.0
pyyaml==5.4.1
pandas==1.3.4
xmltodict==0.12.0
Faker==9.7.1